requirements.txt
flask
mysql-connector-python
boto3